feat: Ability to set the environment variables
The environment variables for the client credentials are expected to be
stored in a secret. The secret can either be created manually or it can
be created by the helm chart.

The environment variable SPOTIFY_PROXY_API_KEY is optional, so if it is
desired by the user, it should also be in a secret (different from the
credentials one). Similarly as above, the secret can either be created
manually or by the helm chart.

The environment variable SPOTIFY_PROXY_BASE_URL is optional, and since
it is not a secret, it can be set directly in the helm chart values.